aquastart/backend/requirements.txt
supabase==2.10.0
python-dotenv==1.0.0
reportlab==4.0.9
requests==2.31.0
python-docx==0.8.11
